Fix function name matching with normalization and improve error handling
- Added NormalizeFunctionName helper to strip UE prefixes (K2_, BP_, EdGraph_, UE_)
- Fixed GetActorLocation matching issue when using class_name parameter
- Removed duplicate Blueprint validation from command layer - service handles it
- Now error messages are properly formatted and readable
- Refactored BlueprintNodeCreationService into 8 modular files
